[SeaBIOS] [PATCH 2/2] bochsvga: Implement vbe 15h function
Posted Wed Mar 7 03:41:07 CET 2012

This patch adds the ability for the Bochs VGA ROM to return information
about a generic SVGA monitor.

It uses a hardcoded EDID for now, but we can easily immagine contructing
an EDID at run time with the help of some support code.

Returning an EDID with decent timings may be necessary in case we want
to use a resolution superior to 1024x768 with Xorg.

Inspired by this patch I digged out from the vgabios mailing list:
http://old.nabble.com/VBE-BIOS-support-for-monitor-information-(EDID)-td22952446.html

Untested with Xorg.

Signed-off-by: Julian Pidancet <julian.pidancet@gmail.com>